READMI added

+++ srcs/includes/colors.h
@@ -1,10 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   colors.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ehillman <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/20 19:09:45 by ehillman          #+#    #+#             */
+/*   Updated: 2021/03/20 19:09:48 by ehillman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef COLORS_H
 # define COLORS_H
 
 typedef struct	s_color {
-	double		r;
-	double		g;
-	double		b;
+	float		r;
+	float		g;
+	float		b;
 }				t_color;
 
 typedef struct	s_phong
@@ -14,6 +26,6 @@ typedef struct	s_phong
 	t_vector	view_dir;
 	t_vector	halfway_dir;
 	t_color		specular;
-	double		spec;
+	float		spec;
 }				t_phong;
 #endif
