+++ srcs/includes/intersect.h
@@ -1,21 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   intersect.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ehillman <ehillman@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/20 19:10:07 by ehillman          #+#    #+#             */
+/*   Updated: 2021/03/21 00:34:18 by ehillman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef MINIRT_INTERSECT_H
 # define MINIRT_INTERSECT_H
 
 void			sphere_start(t_scene *scene, t_sphere *sp,
-				double *min, t_ray ray, t_color *c_tmp);
+				float *min, t_ray ray, t_color *c_tmp);
 void			plane_start(t_scene *scene, t_plane *pl,
-				double *min, t_ray ray, t_color *c_tmp);
+				float *min, t_ray ray, t_color *c_tmp);
 void			triangle_start(t_scene *scene, t_triangle
-				*tr, double *min, t_ray ray, t_color *c_tmp);
+				*tr, float *min, t_ray ray, t_color *c_tmp);
 void			sq_start(t_scene *scene, t_square *sq,
-				double *min, t_ray ray, t_color *c_tmp);
+				float *min, t_ray ray, t_color *c_tmp);
 void			cy_start(t_scene *scene, t_cylinder *cy,
-				double *min, t_ray ray, t_color *c_tmp);
-t_vector		find_cy_normal(double intersec, t_cylinder cy, t_ray ray);
-double			sphere_intersect(t_ray ray, t_sphere *sp);
-double			plane_intersect(t_ray ray, t_plane *plane);
-double			triangle_intersec(t_ray ray, t_triangle *triangle);
-double			square_intersec(t_ray ray, t_square *sq, double min_t);
-double			cy_intersect(t_ray ray, t_cylinder *cy);
+				float *min, t_ray ray, t_color *c_tmp);
+t_vector		find_cy_normal(float intersec, t_cylinder cy, t_ray ray);
+float			sphere_intersect(t_ray ray, t_sphere *sp);
+float			plane_intersect(t_ray ray, t_plane *plane);
+float			triangle_intersec(t_ray ray, t_triangle *triangle);
+float			square_intersec(t_ray ray, t_square *sq, float min_t);
+float			cy_intersect(t_ray ray, t_cylinder *cy);
 
 #endif
