+++ srcs/includes/lists.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   lists.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ehillman <ehillman@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/20 19:10:16 by ehillman          #+#    #+#             */
+/*   Updated: 2021/03/20 19:19:19 by ehillman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef LISTS_H
 # define LISTS_H
 
@@ -10,7 +22,7 @@ typedef struct		s_cameras{
 	t_vector		coordinates;
 	t_vector		direction;
 	void			*next;
-	double			field_of_v;
+	float			field_of_v;
 }					t_cameras;
 
 typedef struct		s_lights {
@@ -18,21 +30,21 @@ typedef struct		s_lights {
 	t_vector		light_dir;
 	void			*next;
 	t_color			color;
-	double			intensity;
+	float			intensity;
 }					t_lights;
 
 typedef struct		s_ab_light {
 	t_color			color;
-	double			intensity;
+	float			intensity;
 }					t_ab_light;
 
 void				push_back_cam(t_cameras *list, t_cameras *cam);
 t_figures			*new_figur_list(void *content, char spec);
 t_cameras			*new_camera_node(t_vector coordinates,
-					t_vector dir, double fov);
+					t_vector dir, float fov);
 t_lights			*new_light_node(t_vector coordinates,
-					double intensity, t_color color);
+					float intensity, t_color color);
 void				push_back_light(t_lights *list, t_vector coordinates,
-					double intensity, t_color color);
+					float intensity, t_color color);
 
 #endif
