+++ srcs/lists_funcs/list_camera_funcs.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   list_camera_funcs.c                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ehillman <ehillman@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/20 19:27:23 by ehillman          #+#    #+#             */
+/*   Updated: 2021/03/20 19:47:16 by ehillman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../includes/MiniRT.h"
 
-t_cameras		*new_camera_node(t_vector coordinates, t_vector dir, double fov)
+t_cameras		*new_camera_node(t_vector coordinates, t_vector dir, float fov)
 {
 	t_cameras	*new;
