+++ srcs/vectors_funcs/vector_normalise.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   vector_normalise.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ehillman <ehillman@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/20 19:37:49 by ehillman          #+#    #+#             */
+/*   Updated: 2021/03/20 19:38:32 by ehillman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../includes/MiniRT.h"
 
 t_vector		vector_normalise(t_vector v)
 {
-	double		len;
-	double		len_coeff;
+	float		len;
+	float		len_coeff;
 	t_vector	n;
 
 	n = new_vector(0, 0, 0);
